refactor: 패키지 레이어 구조 수정
