+++ src/main/java/com/example/live_backend/LiveBackendApplication.java
@@ -2,14 +2,8 @@ package com.example.live_backend;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cache.annotation.EnableCaching;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @SpringBootApplication
-@EnableJpaAuditing
-@EnableFeignClients
-@EnableCaching
 public class LiveBackendApplication {
 	public static void main(String[] args) {
 		SpringApplication.run(LiveBackendApplication.class, args);
