+++ src/main/java/com/example/live_backend/domain/place/config/PlaceWebConfig.java
@@ -1,0 +1,20 @@
+package com.example.live_backend.domain.place.config;
+
+import com.example.live_backend.domain.place.interceptor.RateLimitInterceptor;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+@RequiredArgsConstructor
+public class PlaceWebConfig implements WebMvcConfigurer {
+    
+    private final RateLimitInterceptor rateLimitInterceptor;
+    
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(rateLimitInterceptor)
+            .addPathPatterns("/v1/places/**");
+    }
+}
