+++ src/main/java/com/example/live_backend/domain/place/dto/PlaceItem.java
@@ -1,0 +1,39 @@
+package com.example.live_backend.domain.place.dto;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@Schema(description = "장소 정보")
+public class PlaceItem {
+    
+    @Schema(description = "장소 ID", example = "kakao:123456789")
+    private String id;
+    
+    @Schema(description = "장소명", example = "더마음의원")
+    private String name;
+    
+    @Schema(description = "카테고리 정보")
+    private Category category;
+    
+    @Schema(description = "주소 정보")
+    private Address address;
+    
+    @Schema(description = "위치 좌표")
+    private Location location;
+    
+    @Schema(description = "전화번호", example = "02-1234-5678")
+    private String phone;
+    
+    @Schema(description = "썸네일 URL", example = "https://...")
+    private String thumbnailUrl;
+    
+    @Schema(description = "데이터 소스", example = "kakao")
+    private String source;
+}
