--- src/main/java/com/example/live_backend/domain/places/config/PlaceWebConfig.java
@@ -1,20 +1,0 @@
-package com.example.live_backend.domain.places.config;
-
-import com.example.live_backend.domain.places.interceptor.RateLimitInterceptor;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-@RequiredArgsConstructor
-public class PlaceWebConfig implements WebMvcConfigurer {
-    
-    private final RateLimitInterceptor rateLimitInterceptor;
-    
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(rateLimitInterceptor)
-            .addPathPatterns("/v1/places/**");
-    }
-}
