+++ src/main/java/com/example/live_backend/infra/kakao/feign/KakaoFeignConfig.java
@@ -4,12 +4,14 @@ import feign.Logger;
 import feign.Request;
 import feign.RequestInterceptor;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 import java.util.concurrent.TimeUnit;
 
 @Configuration
+@EnableFeignClients
 public class KakaoFeignConfig {
 
     @Value("${kakao.local.rest-api-key}")
